% Generated by roxygen2 (4.0.0): do not edit by hand
\name{countPointsInGrid}
\alias{countPointsInGrid}
\title{Count number of points in each cell of a spatial grid}
\usage{
countPointsInGrid(xy.points, xy.centers = NULL, template = NULL)
}
\arguments{
\item{xy.points}{}

\item{xy.centers}{}

\item{template}{}
}
\value{
Describe the value returned If it is a LIST, use
\item{comp1 }{Description of 'comp1'}
\item{comp2 }{Description of comp2'}
}
\description{
Count number of points in each cell of a spatial grid defined by controid points
}
\details{
There are no details
}
\author{
Devin S. Johnson
}

